Modularize query logic into a hook and simplify
