+++ client/my-sites/stats/stats-subscribers-overview/index.tsx
@@ -1,86 +1,24 @@
 import { CountComparisonCard } from '@automattic/components';
-import { translate } from 'i18n-calypso';
 import React from 'react';
-import {
-	useSubscribersQueries,
-	SubscribersData,
-} from 'calypso/my-sites/stats/hooks/use-subscribers-query';
-
-// array of indices to use to calculate the dates to query for
-const cardIndices = [ 0, 30, 60, 90 ];
+import useSubscribersOverview from 'calypso/my-sites/stats/hooks/use-subscribers-overview';
 
 interface SubscribersOverviewProps {
 	siteId: number | null;
 }
 
-// calculate the date to query for based on the number of days to subtract
-function calculateQueryDate( daysToSubtract: number ) {
-	const today = new Date();
-	const date = new Date( today );
-	date.setDate( date.getDate() - daysToSubtract );
-	return date.toISOString().split( 'T' )[ 0 ];
-}
-
-// calculate the stats to display in the cards
-function SubscribersOverviewCardStats( subscribersData: SubscribersData, index: number ) {
-	const getCount = () => {
-		return subscribersData?.data?.[ 0 ]?.subscribers || 0;
-	};
-
-	let heading;
-	switch ( index ) {
-		case 0:
-			heading = translate( 'Today' );
-			break;
-		case 1:
-			heading = translate( '30 days ago' );
-			break;
-		case 2:
-			heading = translate( '60 days ago' );
-			break;
-		case 3:
-			heading = translate( '90 days ago' );
-			break;
-		default:
-			heading = '';
-			break;
-	}
-
-	const overviewCardStat = {
-		heading,
-		count: getCount(),
-	};
-
-	return overviewCardStat;
-}
-
 const SubscribersOverview: React.FC< SubscribersOverviewProps > = ( { siteId } ) => {
-	const period = 'day';
-	const quantity = 1;
-	const dates = cardIndices.map( calculateQueryDate );
-
-	const subscribersQueries = useSubscribersQueries( siteId, period, quantity, dates );
+	const { isLoading, isError, overviewData } = useSubscribersOverview( siteId );
 
 	return (
 		<div className="subscribers-overview highlight-cards">
 			<div className="highlight-cards-list">
-				{ subscribersQueries.map( ( result, index ) => {
-					if ( result.isLoading ) {
-						return <div key={ index }>Loading...</div>;
-					}
-
-					if ( result.isError ) {
-						return <div key={ index }>Error: Failed to load data.</div>;
-					}
-
-					const subscribersData = result.data;
-					const overviewCardStat = SubscribersOverviewCardStats( subscribersData, index );
-
+				{ overviewData.map( ( { count, label }, index ) => {
 					return (
+						// TODO: Communicate loading vs error state to the user.
 						<CountComparisonCard
-							key={ overviewCardStat.heading }
-							heading={ overviewCardStat.heading }
-							count={ overviewCardStat.count }
+							key={ index }
+							heading={ label }
+							count={ isLoading || isError ? null : count }
 							showValueTooltip
 							icon={ false }
 						/>
